Add ready-to-use TogetherAI backend
TogetherAI is a service similar to OpenAI's, but they give you a free
trial with 25$ starting credit, which IMHO makes it ideal to try out
Gptel.

They use open-source (I think) models which are not as SOTA as OpenAI,
but OTOH seem to have read somewhere they're significantly cheaper to
access.

* gptel.el (gptel--togetherai): New variable.
(gptel-backend): Add togetherai as an option
(gptel): Check if gpt-model defcustom is actually available